Finish coding get_dynamics_constraint function

+++ cpp/linear_mpc/include/linear_mpc/matrix_algebra.h
@@ -33,25 +33,22 @@ Eigen::MatrixXd kron(const Eigen::MatrixXd &m1, const Eigen::MatrixXd &m2) {
 }
 
 /**
- * @brief Create a block diagonal matrix from provided matrices
- *        3 input version
- *        inspired scipy.linalg.block_diag
- *        https://docs.scipy.org/doc/scipy-0.14.0/reference/generated/scipy.linalg.block_diag.html
+ * @brief Create a block diagonal matrix from provided matrices 3 input version
  *
  * @param m1 first matrix
  * @param m2 second matrix
  *
  * @return Created a block diagonal matrix
  */
-Eigen::MatrixXf block_diag(const Eigen::MatrixXf &m1,
-                           const Eigen::MatrixXf &m2) {
+Eigen::MatrixXd block_diag(const Eigen::MatrixXd &m1,
+                           const Eigen::MatrixXd &m2) {
 
   uint32_t m1r = m1.rows();
   uint32_t m1c = m1.cols();
   uint32_t m2r = m2.rows();
   uint32_t m2c = m2.cols();
 
-  Eigen::MatrixXf mf = Eigen::MatrixXf::Zero(m1r + m2r, m1c + m2c);
+  Eigen::MatrixXd mf = Eigen::MatrixXd::Zero(m1r + m2r, m1c + m2c);
   mf.block(0, 0, m1r, m1c) = m1;
   mf.block(m1r, m1c, m2r, m2c) = m2;
